+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/search_box/SearchEngineIntegrationTest.kt
@@ -661,8 +661,8 @@ internal class SearchEngineIntegrationTest : BaseTest() {
                 OpenHours.Scheduled(
                     periods = listOf(
                         OpenPeriod(
-                            open = WeekTimestamp(WeekDay.SATURDAY, 7, 0),
-                            closed = WeekTimestamp(WeekDay.SATURDAY, 20, 0)
+                            open = WeekTimestamp(WeekDay.FRIDAY, 7, 0),
+                            closed = WeekTimestamp(WeekDay.FRIDAY, 20, 0)
                         )
                     )
                 ), openHours
@@ -957,7 +957,7 @@ internal class SearchEngineIntegrationTest : BaseTest() {
         searchEngine.retrieve(mapboxId, callback)
 
         val result = callback.getResultBlocking()
-        assertTrue(result is Exception)
+        assertTrue(result.isError)
     }
 
     @Test