+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog for the Mapbox Search SDK for Android
 
+## 2.6.1
+
+### New features
+- Added experimental Details API that provides access to POI metadata, boundary details, addresses and places. See `DetailsApi` type.
+
+### Bug fixes
+- Fixed `OpenHours` parsing for the Search Box Api type.
+
+### Mapbox dependencies
+- Search Native SDK `2.6.2`
+- Common SDK `24.8.0`
+
+
+
 ## 2.6.0
 
 ### Mapbox dependencies